Move logic to a 'populate' service, to allow a new populate_all_collectivities script
