+++ src/geofr/services/__init__.py
@@ -1,0 +1,5 @@
+"""Perimeter definition and management.
+
+Also contains the tasks to update the different perimeters (regions,
+departments, cities, etc.)
+"""
